test: move preview smoke code into package
