+++ scripts/preview_smoke/__init__.py
@@ -1,0 +1,5 @@
+"""Preview-deployment smoke runner package."""
+
+from scripts.preview_smoke.runner import main, run_smoke
+
+__all__ = ["main", "run_smoke"]
